requirements.txt
numpy
pandas
scikit-learn
matplotlib
pyarrow
pyyaml